requirements.txt
python-jenkins==1.8.0
python-dotenv
pandas